# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PageGuard REST API
  version: "1.0.0"
  description: |
    Versioned REST API for PageGuard, parallel to the browser dashboard.
    Bearer-token authenticated via the `API_TOKEN` server env var.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/urls:
    get:
      summary: List monitored URLs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 1000, default: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        "200":
          description: Paginated URL list
          content:
            application/json:
              schema:
                type: object
                required: [urls, total]
                additionalProperties: false
                properties:
                  urls:
                    type: array
                    items: { $ref: "#/components/schemas/MonitoredUrl" }
                  total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TokenNotConfigured" }
    post:
      summary: Add a monitored URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, label]
              additionalProperties: false
              properties:
                url:
                  type: string
                  format: uri
                  description: http(s) URL, SSRF-validated against private/reserved addresses
                label:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MonitoredUrl" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/TokenNotConfigured" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Bad input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    TokenNotConfigured:
      description: API_TOKEN env var is not configured on the server
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error: { type: string }
        field: { type: string }
...
    MonitoredUrl:
      type: object
      additionalProperties: false
      required:
        - id
        - url
        - label
        - url_hash
        - status
        - last_checked
        - last_change
        - reference_capture_id
        - created_at
      properties:
        id: { type: integer }
        url: { type: string }
        label: { type: string }
        url_hash: { type: string }
        status:
          type: string
          enum: [ok, change_detected, error, pending]
        last_checked:
          oneOf:
            - type: "null"
            - type: string
        last_change:
          oneOf:
            - type: "null"
            - type: string
        reference_capture_id:
          oneOf:
            - type: "null"
            - type: integer
        created_at: { type: string }
...